ch(refactor): update tests and configuration
