+++ server/controllers/welcome.controllers.js
@@ -1,0 +1,5 @@
+export default {
+  welcome(req, res) {
+    res.status(200).send({ success: true, message: 'Welcome to Ballers API endpoint' });
+  },
+};
