+++ server/routes/welcome.routes.js
@@ -1,9 +1,19 @@
 import express from 'express';
+import WelcomeController from '../controllers/welcome.controllers';
 
 const router = express.Router();
 
-router.get('/', (req, res) => {
-  res.status(200).send({ success: true, message: 'Welcome to Ballers API endpoint' });
-});
+/**
+ * @swagger
+ * path:
+ *  /:
+ *    get:
+ *      summary: A welcome text from Ballers API
+ *      tags: [Welcome]
+ *      responses:
+ *        "200":
+ *          description: A welcome object
+ */
+router.get('/', WelcomeController.welcome);
 
 module.exports = router;
