+++ test/controllers/welcome.controller.test.js
@@ -3,7 +3,7 @@ import { expect, request } from '../config';
 describe('Welcome Route', () => {
   it('should have status 200', (done) => {
     request()
-      .get('/api/v1/welcome')
+      .get('/api/v1/')
       .end((err, res) => {
         expect(res).to.have.status(200);
         done();
@@ -12,7 +12,7 @@ describe('Welcome Route', () => {
 
   it('should return a valid object', (done) => {
     request()
-      .get('/api/v1/welcome')
+      .get('/api/v1/')
       .end((err, res) => {
         expect(res.body).to.eql({
           success: true,
